backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["mihett05 <mihett05@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
SQLAlchemy = "^1.4.46"
alembic = "^1.9.1"
poetry = "^1.3.1"
pyjwt = "^2.6.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
